requirements.txt
streamlit==1.37.1
openai==1.40.3
python-dotenv==1.0.1
pydub==0.25.1
transformers==4.44.2
torch>=2.0.0
torchaudio>=2.0.0
scipy==1.14.1
numpy==1.26.4
pandas==2.2.2
pytest==8.3.2
